fixed winning player display name and stage the code to refactor
