+++ lib/menu.rb
@@ -1,4 +1,4 @@
-require_relative './helper.rb'
+require_relative './input_validator_helper.rb'
 module MenuHelper
   include ValidatorHelper
   def table(board)
@@ -14,6 +14,7 @@ module MenuHelper
   end
 
   def player_one_info(player)
+    welcome_message
     puts 'Player 1 - Please enter your name:'
     player.username_one = gets.chomp
 
@@ -27,7 +28,7 @@ module MenuHelper
         puts 'Okay, good name!'
       else
         puts "Your name is invalid again. We'll just call you 'Player 1'"
-        player.username_one = 'Player 1'.freeze
+        player.username_one = 'Player 1'
       end
     end
     puts ''
@@ -38,7 +39,6 @@ module MenuHelper
     player.username_two = gets.chomp
 
     if validate_name(player.username_two)
-
     else
       puts 'Sorry! Invalid name. Try again.'
       puts 'Player 2 - please enter your name: '
@@ -47,15 +47,15 @@ module MenuHelper
         puts 'Okay, good name!'
       else
         puts "Your name is invalid again. We'll just call you 'Player 1'"
-        player.username_two = 'Player 2'.freeze
+        player.username_two = 'Player 2'
       end
     end
     puts ''
   end
 
-  def notice(player)
-    puts "#{player.username_one} you will play as X"
-    puts "#{player.username_two} you are gonna be O"
+  def notice(players)
+    puts "#{players.username_one} you will play as X"
+    puts "#{players.username_two} you are gonna be O"
     puts ''
   end
 end
